requirements.txt
fastapi[standard]
torch
torchvision
timm
python-multipart
pillow